Finish file-per-short style Installation repos
After some deliberation, I decided to at least initially attempt a file
model for each short-id rather than a directory model.  First, I noticed
that I would need to add a repository hash to the Distribution class for
uninstall to properly locate the file internal to each short-id's folder.
Second, what to use for each internal file's name? Could be a sha1 of
the long spec but then it just seems redundant as the file's contents
are completely unneeded.

So instead of that, I went for a one-file model: Installation's
install appends the long spec to the file and uninstall just matches
the dist.Str against the lines deleting the first match. need-repository
locates the short-id file and matches the rspec against each line. Filesystem's
need-repository searches the META6 for provides-repository hash and
gathers the value of the matching short-id as key.